Move functions to separate file and create routes folder
