+++ app/routes/route-utils.js
@@ -1,0 +1,132 @@
+const faker = require('faker')
+const path = require('path')
+const moment = require('moment')
+const filters = require('./../filters.js')()
+const _ = require('lodash')
+
+
+// Return first part of url to use in redirects
+exports.getRecordPath = req => {
+  let recordType = req.params.recordtype
+  return (recordType == 'record') ? (`/record/${req.params.uuid}`) : '/new-record'
+}
+
+// Delete temporary stores of data
+exports.deleteTempData = (data) => {
+  delete data.degreeTemp
+  delete data.record
+}
+
+exports.addEvent = (record, content) => {
+  record.events.items.push({
+    title: content,
+    user: 'Provider',
+    date: new Date()
+  })
+}
+
+// Append referrer to string if it exists
+exports.getReferrer = referrer => {
+  if (referrer && referrer != 'undefined'){
+    return `?referrer=${referrer}`
+  }
+  else return '' 
+}
+
+// Stolen from manage
+exports.getTimeline = (record) => {
+  return record.events.items.map(item => {
+    return {
+      label: {
+        text: item.title
+      },
+      datetime: {
+        timestamp: item.date,
+        type: 'datetime'
+      },
+      byline: {
+        text: item.user
+      }
+      // link: getLink(item, record)
+    }
+  }).reverse()
+}
+
+
+// Check if all sections are complete
+exports.recordIsComplete = record => {
+  if (!record) return false
+  let regularSections = [
+    'personalDetails',
+    'contactDetails',
+    'diversity',
+    'programmeDetails',
+    // 'qualifications.gceStatus', // A levels not implimented yet
+    'gcse',
+    'degree'
+  ]
+
+  let recordIsComplete = true
+
+  regularSections.forEach(section => {
+    if (_.get(record, `${section}.status`) != "Completed"){
+      recordIsComplete = false
+    }
+  })
+
+  return recordIsComplete
+}
+
+// Update or create a record
+exports.updateRecord = (data, newRecord, timelineMessage) => {
+
+  if (!newRecord) return false
+  
+  let records = data.records
+  newRecord.updatedDate = new Date()
+  if (timelineMessage !== false){
+    let message = (timelineMessage) ? timelineMessage : "Record updated"
+    exports.addEvent(newRecord, message)
+  }
+  if (newRecord.addressType == "domestic"){
+    delete newRecord.contactDetails.internationalAddress
+  }
+  if (newRecord.addressType == "international"){
+    delete newRecord.contactDetails.address
+  }
+  data.record = newRecord
+
+  if (newRecord.personalDetails && !newRecord.personalDetails.fullName){
+    Object.defineProperty(newRecord.personalDetails, 'fullName', {
+      get() {
+        let names = []
+        names.push(this.givenName)
+        names.push(this.middleNames)
+        names.push(this.familyName)
+        return names.filter(Boolean).join(' ')
+      },
+      enumerable: true
+    })
+  }
+  if (newRecord.personalDetails && !newRecord.personalDetails.shortName){
+    Object.defineProperty(newRecord.personalDetails, 'shortName', {
+      get() {
+        let names = []
+        names.push(this.givenName)
+        names.push(this.familyName)
+        return names.filter(Boolean).join(' ')
+      },
+      enumerable: true
+    })
+  }
+
+  if (!newRecord.id){
+    newRecord.id = faker.random.uuid()
+    records.push(newRecord)
+  }
+  else {
+    let recordIndex = records.findIndex(record => record.id == newRecord.id)
+    records[recordIndex] = newRecord
+  }
+  return true
+}
